update from brazil
updated comments for global variables

updated the outer ring/inner equalizer bars to not have to be of the
same enumeration.

updated bars/levels ranges so it is capped at 100, but in practice it
seems to peak at 50, will need some more scaling code added here to
scale up during soft parts of songs.